fix: :bug: Viewing the calendar in any timezone other than EST would cause a fatal error.
A special thanks to Bryce and Edward for showing me this error! The issue was related how the dates were being stores in the DB as well as a bug I also found in the getDateWeek Function, I borrowed from someone online. I wrote my own solution to account daylight savings, which resolved the issue.